test(skill): isolate runtime-effect regressions
